Handle unique by db name in status
